backend/requirements.txt
fastapi[standard]>=0.115.0
google-genai>=1.0.0
sse-starlette>=2.0.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
